refactor: refactor LCDSeminar and OSBSeminar as individual components
